requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
youtube-transcript-api>=0.6.1
pydantic>=2.4.2
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
requests>=2.31.0